+++ ckanext/storepublisher/templates/package/publish.html
@@ -13,7 +13,7 @@
          FIWARE Store is the main FIWARE showcase and is the best way to sell/advertise your dataset. This form ease you the task of creaing an offering in the FIWARE Store. Just fill the following fields. Some fields are filled automatically based on the dataset but you can change their content.
       </p>
       <p>
-         By publishing the dataset in the FIWARE Store, you will control the user that can access the dataset. Every time a user acquire the dataset, the list of allowed users will be updated automatically.
+         By publishing the dataset in the FIWARE Store, you will control the user that can access the dataset. Every time a user acquire the dataset, the list of allowed users will be updated automatically. Additionally, you can either charge users or force them to accept some legal terms for using your dataset.
       </p>
     </div>
   </section>
